Moving RemoveEditorViewTests to its own file
